# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Envoitures Rest Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get List of Reviews",
                "operationId": "346169ce6ab0c633530a7cedc1a25fab",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "userId": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "rideId": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "isPrivate": {
                                                        "type": "number",
                                                        "example": "2"
                                                    },
                                                    "note": {
                                                        "type": "number",
                                                        "example": "4"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "This ride was absolutely fabulouse"
                                                    },
                                                    "updatedAt": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "createdAt": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create Review",
                "operationId": "154c71f5f72b459d998f5ea451bd4c96",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "userId": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "rideId": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "isPrivate": {
                                                "type": "number",
                                                "example": "2"
                                            },
                                            "note": {
                                                "type": "number",
                                                "example": "4"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "This ride was absolutely fabulouse"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "userId": "2",
                                    "rideId": "2",
                                    "isPrivate": "0",
                                    "note": "2",
                                    "content": "This ride was absolutely fabulouse"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "userId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "rideId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "isPrivate": {
                                            "type": "number",
                                            "example": "2"
                                        },
                                        "note": {
                                            "type": "number",
                                            "example": "4"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "This ride was absolutely fabulouse"
                                        },
                                        "updatedAt": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/{id}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get Review Details",
                "operationId": "a9949d3a965ed93ae30547c07eef7505",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "userId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "rideId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "isPrivate": {
                                            "type": "number",
                                            "example": "2"
                                        },
                                        "note": {
                                            "type": "number",
                                            "example": "4"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "This ride was absolutely fabulouse"
                                        },
                                        "updatedAt": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update Review",
                "operationId": "17776becb9faf054b5c49bc83ad6c974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "userId": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "rideId": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "isPrivate": {
                                                "type": "number",
                                                "example": "2"
                                            },
                                            "note": {
                                                "type": "number",
                                                "example": "4"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "This ride was absolutely fabulouse"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "userId": "2",
                                    "rideId": "2",
                                    "isPrivate": "0",
                                    "note": "2",
                                    "content": "This ride was absolutely fabulouse"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "userId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "rideId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "isPrivate": {
                                            "type": "number",
                                            "example": "2"
                                        },
                                        "note": {
                                            "type": "number",
                                            "example": "4"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "This ride was absolutely fabulouse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete Review",
                "operationId": "89986343b5dbd2b909c2764aeb7e4096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete review success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}